move password hasher in dedicated file
